requirements.txt
Flask>=2.0.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
